requirements.txt
google-generativeai
flask
firebase-admin